Cleaning up version conflict between Changelog and Git tags
Looking at the repository history an merge was incorrectly applied to this project from another. This lead to a confusion over which version the project was at and lead to a mis-match between the project Changelog (which is correct symantically) and the tags within the repository.

As it isn't easy to revoke the incorrectly assigned Git tags (and we should do so anyway) it is simplier to add a relevant message in the Changelog and carry on from the "tagged" versions.